API
Add api support for 5 days weather.

+++ app/src/main/java/com/example/homeassistant/domain/api/dto/LocationDto.kt
@@ -4,7 +4,7 @@ import com.squareup.moshi.Json
 
 data class LocationDto (
    @Json(name = "lon")
-    val longitude: Double? = null,
+    val longitude: Float? = null,
    @Json(name = "lat")
-    val latitude: Double? = null
+    val latitude: Float? = null
 )
